Adding shared XAML for Scenario 1 & 2
Removing the 'blank' XAML and adding references to the UserInfo\shared XAML. The code behind classes contain stubs for all of the event handlers and methods/properties the XAML binds to.

The Scenario2_WatchUsers::Users property stub may not be implemented correctly. It just returns a vector, it may need to return an observable vector.  See:

https://docs.microsoft.com/en-us/windows/uwp/cpp-and-winrt-apis/binding-collection