Fixing issues with populating ScenarioControl
The MainPage defined method that OnNavigateTo() used to get the scenarios was returning a StringMap which is unordered. Therefore the scenarios were not added in the order they were put in the map.

Switching the method to a std::vector. Since only C++ code will be using these methods there is no need to make the datatypes WinRT compatible/exposable.