First attempt at populating UserList in Scenario1
The C++/CX version would create & add the model for the list box (a vector) then append items to it asynchronously. In this attempt I'm creating model, then adding the model to the list box when all of the items are ready. This isn't working, the list box is always empty.

I think the model needs to be an observable vector.