Adding observable vector for the Scenario1 model
Adding an implementation I found on the C++/WinRT 'official' docs site. When using it the app crashes.

If I follow the pattern of the C++/CX app where it sets the model for the list box, then adds the items to the observable vector it crashes in a 'deep' stack, not sure exactly the cause.

If I follow the pattern I attempted for the regular vector, where it builds the vector then sets the model, it crashes while building the vector in a template of Append().