Fixing construction of observable vector
With these changes the list box is still not being populated, but it is no longer crashing.

Using the correct 'technique' to construct the observable vector fixed the crash. It needed to be constructed by winrt::make().

Also changing the observable vector implementation to remove the "using namespace". This was not causing any problems but changing it this way made it more like the implementation of input_vector I found in winrt/base.h.