Refactoring Scenario1 GetUsersAsync()
In trying to determine why the UserViewModel is not being displayed in the UserList I read that the bindable objects must be created on the UI thread. I am refactoring this method to be more like the C++/CX version, even though that app's implementation did not attempt to move any of the calls back to the UI thread (even the call the UserList.SelectedIndex).