Testing XAML binding in Scenario1
Adding some Debug logging to check when the list of UserViewModels (models) is changed. Also adding a new button and binding in the XAML to a single UserViewModel. There is debug logging watching for modifications to this also.

At this time the XAML binding is not working correctly. When new items are added to models blank items show up in the combox box. The new button shows the bound value when it is initially displayed. But when the model behind it changes, it doe snot update. Yet, both of my 'watchers' fire and I see the debug logging.