Switching UserInfoCppWinRt to Win 10 1809
Switching this project to use the Windows Platform SDK 10.0.17763.0 which is Win 10 1809. Leaving the min version at 10.0.17134.0 which is Win 10 1803.

The new SDK required a couple of project build changes and including more headers in the pre-compiled header file.

Also switching back to 'Comformance mode', adding back the /permissive- switch.