Fixing one issue with the test controls
One of the tests added in this branch was a button that was bound to _currentUser.DisplayName. When "Show Properties" is clicked it sets the display name to "Changed!".

To get this to work we must modify the XAML for the test button, adding "Mode=OneWay" to the binding expression. I found this in the MS docs on C++/WinRT, but there is no reason given why we have to do this.

https://docs.microsoft.com/en-us/windows/uwp/cpp-and-winrt-apis/binding-property