Changing the type of the userViewModels
According to the C++/WinRT doc on binding to collections, the type of the collection should be IObservableVector<SDKTemplate::UserViewModel>. I tried this and the binding didn't work, there were the correct number of entries but they were all blank.

So, switching to just IObservable<IInspectable>. Also doing some renaming and fixed a bug. In the for loop that copies the displayNames to the models it was always copying displayNames[0].

https://docs.microsoft.com/en-us/windows/uwp/cpp-and-winrt-apis/binding-collection