Fixing issue with ListBox displaying empty entries
This is the same issue with the ListBox on Scenario2_WatchUsers that the ComboBox on Scenario1_FindUsers had.

The DisplayMemberPath attribute on the ListBox does not work on C++/WinRT. The solution is to more explicitly tell the control about the items in the supplied list.

Also changing the type of the vector of UserViewModels (as 'recommended' in the doc below) to IObservableVector<>.

https://docs.microsoft.com/en-us/windows/uwp/cpp-and-winrt-apis/binding-collection