Refactoring use of GetDisplayNameOrGenericNameAsync()
Both scenarios had the same method to get the display name or a made up name that included a user count. The main difference was how the two implementations kept track of the user count.

Made a version of this function in Scenario1_FindUsers.cpp that can be shared by both.