Getting UserInfo/cppwinrt/UserInfoCppWinRt.vcxproj compiling after rebase
I rebased this branch on master (two years after it was created). It no longer includes the SharedContent/cppwinrt/App.cpp and SharedContent/cppwinrt/MainPage.cpp that I wrote, it uses the ones in master (master now has C++/WinRT samples).

The biggest changes required for this switch, code-wise, were to SampleConfiguration.cpp/.h because it needs to "interact" with the new App & MainPage. The ways that the Scenario pages interface with MainPage still work. They just needed to include one generated file to satisfy a linker requirement.

Also, with the earlier upgrade to the latest Windows SDK, we are able to get rid of the implementation of single_threaded_observable_vector. It was copied from a MS doc page long ago before there was an implementation included in the SDK.

Also changing the .vcxproj file to be more like the other C++/WinRT samples.